Move things around for clarity (#101)
* move things around for clarity

* rename proto/private to proto/private_key
