+++ src/proto/messaging.proto
@@ -96,29 +96,3 @@ message Message {
     V1Message v1 = 1;
   }
 }
-
-// Private Key Storage
-//
-// Following definitions are not used in the protocol, instead
-// they provide a way for encoding private keys for storage.
-
-message PrivateKey {
-  message Secp256k1 {
-      bytes bytes = 1; // D big-endian, 32 bytes
-  }
-  uint64 timestamp = 1;
-  oneof union {
-      Secp256k1 secp256k1 = 2;
-  }
-  PublicKey publicKey = 3;
-}
-
-message PrivateKeyBundle {
-  PrivateKey identityKey = 1;
-  repeated PrivateKey preKeys = 2;
-}
-
-message EncryptedPrivateKeyBundle {
-  bytes walletPreKey = 1;
-  Ciphertext ciphertext = 2;
-}
